remove old solver
